sandbox/python/py_pydantic_ai/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langgraph>=0.4.0",
    "pydantic-ai>=1.59.0",
]

[tool.uv.workspace]
members = ["bla"]
